refactor : env 관련 변수 설정 추가, 리프레쉬 토큰 관련 entity column 추가, utiles 폴더명 변경 - env 값 local, prod, dev 구분하여 추가 : ci/cd 과정에서 각 변수를 env 값으로 활용하면 될 것 같습니다. - refresh token이 저장될 컬럼(refresh_token)을 user 테이블에 추가했습니다. - 계속 신경쓰였던 utiles 폴더 이름을 utils로 변경
