+++ .env.sample
@@ -1,11 +1,28 @@
 PORT=포트번호
 
 DB_TYPE=mysql
-DB_HOST=ip주소
 DB_PORT=db포트번호
 DB_USERNAME=db계정
-DB_PASSWORD=db비밀번호
-DB_DATABASE=db이름
+
+DB_HOST_PROD=RDS_ip주소
+DB_HOST_DEV=RDS_ip주소
+DB_HOST_LOCAL=local_ip주소
+
+DB_PASSWORD_PROD=PROD_db비밀번호
+DB_PASSWORD_DEV=DEV_db비밀번호
+DB_PASSWORD_LOCAL=local_db비밀번호
+
+DB_DATABASE_PROD=PROD_SERVER_db이름
+DB_DATABASE_DEV=DEV_SERVER_db이름
+DB_DATABASE_LOCAL=LOCAL_SERVER_db이름
+
+# DB_LOGGING_PROD = false
+# DB_LOGGING_DEV = true
+DB_LOGGING_LOCAL = true
+
+# DB_SYNCHRONIZE_PROD = false
+# DB_SYNCHRONIZE_DEV = true
+DB_SYNCHRONIZE_LOCAL = true
 
 JWT_SECRET_KEY = jwt secret key
 JWT_EXPIRES_IN = jwt expires in
