+++ src/community/community.service.ts
@@ -1,7 +1,7 @@
 import { Comment } from './../entities/Comment';
 import { HttpException, Injectable, HttpStatus } from '@nestjs/common';
 import { CommunityRepository } from './community.repository';
-import { uploadToS3, getS3Data, deleteS3Data } from '../utiles/aws';
+import { uploadToS3, getS3Data, deleteS3Data } from '../utils/aws';
 import {
   CreateCommentDto,
   CreateOrDeleteCommentLikesDto,
